DLS-10889: Removed jsonx
 - extraced models into separate files
 - removed unused attributes from NSIGetAccountByNinoResponse
 - update libraries
 - removed jsonx
